*** empty log message ***

+++ web/WEB-INF/src/org/pgist/packages/UserPackageCreatorTunerAction.java
@@ -94,8 +94,8 @@ public class UserPackageCreatorTunerAction extends Action {
                 if (value>max) max = value;
             }
             
-            if (min % 10 > 0) {
-                min = (min / 10 + 1) * 10;
+            if ( ((int) min) % 10 > 0) {
+                min = (((int) min) / 10 + 1) * 10;
             }
             
             request.setAttribute("min", min);
